refactor: add custom base edge
