+++ apps/journeys-admin/src/components/Editor/Slider/JourneyFlow/edges/BaseEdge/index.ts
@@ -1,0 +1,1 @@
+export { BaseEdge } from './BaseEdge'
