+++ apps/journeys-admin/src/components/Editor/Slider/JourneyFlow/edges/StartEdge/StartEdge.tsx
@@ -1,14 +1,14 @@
 import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
 import { useTranslation } from 'next-i18next'
-import { ReactElement, useState } from 'react'
+import { ReactElement } from 'react'
 import {
-  BaseEdge,
   EdgeLabelRenderer,
   EdgeProps,
-  getBezierPath,
-  useOnSelectionChange
+  getBezierPath
 } from 'reactflow'
+
+import { BaseEdge } from '../BaseEdge'
 
 export function StartEdge({
   id,
@@ -21,7 +21,6 @@ export function StartEdge({
   style = {}
 }: EdgeProps): ReactElement {
   const { t } = useTranslation('apps-journeys-admin')
-  const [edgeSelected, setEdgeSelected] = useState(false)
   const [edgePath, labelX, labelY] = getBezierPath({
     sourceX,
     sourceY,
@@ -31,30 +30,9 @@ export function StartEdge({
     targetPosition
   })
 
-  useOnSelectionChange({
-    onChange: (selected) => {
-      const selectedEdge = selected.edges.find((edge) => edge.id === id)
-      if (selectedEdge != null) {
-        setEdgeSelected(true)
-      } else {
-        setEdgeSelected(false)
-      }
-    }
-  })
 
   return (
-    <>
-      <BaseEdge
-        path={edgePath}
-        markerEnd={`url(#1__color=${
-          edgeSelected ? '#C52D3A' : 'lightGrey'
-        }&height=10&type=arrowclosed&width=10)`}
-        style={{
-          strokeWidth: 2,
-          stroke: edgeSelected ? '#C52D3A' : '#0000001A',
-          ...style
-        }}
-      />
+    <BaseEdge id={id} style={style} edgePath={edgePath}>
       <EdgeLabelRenderer>
         <Box
           style={{
@@ -71,6 +49,6 @@ export function StartEdge({
           </Typography>
         </Box>
       </EdgeLabelRenderer>
-    </>
+    </BaseEdge>
   )
 }
